pyproject.toml
[project]
name = "my-sujbot"
version = "0.1.0"
description = "Document extraction and analysis bot with hierarchical structure support"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Core dependencies
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",

    # PDF Processing
    "pypdf2>=3.0.0",

    # IBM Docling - Advanced document structure extraction
    "docling>=2.57.0",
    "docling-core>=2.48.0",
    "docling-parse>=4.4.0",
    "docling-ibm-models>=3.9.0",

    # Docling dependencies
    "pypdfium2>=4.30.0",
    "filetype>=1.2.0",
    "rtree>=1.3.0",
    "python-docx>=1.1.2",
    "python-pptx>=1.0.2",
    "openpyxl>=3.1.5",
    "beautifulsoup4>=4.12.3",
    "pandas>=2.1.4",
    "lxml>=4.0.0",
    "pillow>=10.0.0",
    "pylatexenc>=2.10",
    "scipy>=1.6.0",
    "marko>=2.1.2",

    # ML/AI for Docling
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.34.0",
    "accelerate>=1.0.0",
    "safetensors>=0.4.3",

    # OCR support
    "rapidocr>=3.3.0",
    "opencv-python>=4.5.1",
    "ocrmac>=1.0.0",

    # LLM & Text Processing
    "anthropic>=0.18.0",
    "openai>=1.3.0",
    "langchain-text-splitters>=0.0.1",

    # CLI & Utilities
    "click>=8.1.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.7.0",
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
